requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0

# For Phase 3 AWS integration
# boto3==1.34.0
